v3.0.3

+++ src/main.c
@@ -1,11 +1,15 @@
-/* 17mar10abu
+/* 04jun10abu
  * (c) Software Lab. Alexander Burger
  */
 
 #include "pico.h"
 
+#ifdef __CYGWIN__
+#define O_ASYNC FASYNC
+#endif
+
 /* Globals */
-int Signal, Repl, Chr, Slot, Spkr, Mic, Hear, Tell, Children, ExtN;
+int Repl, Chr, Slot, Spkr, Mic, Hear, Tell, Children, ExtN;
 char **AV, *AV0, *Home;
 child *Child;
 heap *Heaps;
@@ -19,11 +23,14 @@ outFile *OutFile, **OutFiles;
 int (*getBin)(void);
 void (*putBin)(int);
 any TheKey, TheCls, Thrown;
-any Alarm, Line, Zero, One, Intern[IHASH], Transient[IHASH], Extern[EHASH];
+any Alarm, Sigio, Line, Zero, One;
+any Intern[IHASH], Transient[IHASH], Extern[EHASH];
 any ApplyArgs, ApplyBody, DbVal, DbTail;
 any Nil, DB, Meth, Quote, T;
 any Solo, PPid, Pid, At, At2, At3, This, Dbg, Zap, Ext, Scl, Class;
 any Run, Hup, Sig1, Sig2, Up, Err, Msg, Uni, Led, Tsm, Adr, Fork, Bye;
+bool Break;
+sig_atomic_t Signal[SIGIO+1];
 
 static int TtyPid;
 static word2 USec;
@@ -79,32 +86,41 @@ void sighandler(any ex) {
 
    if (!Env.protect) {
       Env.protect = 1;
-      switch (Signal) {
-      case SIGHUP:
-         Signal = 0,  run(val(Hup));
-         break;
-      case SIGINT:
-         Signal = 0;
-         if (Repl < 2)
-            brkLoad(ex ?: Nil);
-         break;
-      case SIGUSR1:
-         Signal = 0,  run(val(Sig1));
-         break;
-      case SIGUSR2:
-         Signal = 0,  run(val(Sig2));
-         break;
-      case SIGALRM:
-         Signal = 0,  run(Alarm);
-         break;
-      case SIGTERM:
-         for (flg = NO, i = 0; i < Children; ++i)
-            if (Child[i].pid  &&  kill(Child[i].pid, SIGTERM) == 0)
-               flg = YES;
-         if (!flg)
-            Signal = 0,  bye(0);
-         break;
-      }
+      do {
+         if (Signal[SIGIO]) {
+            --Signal[0], --Signal[SIGIO];
+            run(Sigio);
+         }
+         else if (Signal[SIGUSR1]) {
+            --Signal[0], --Signal[SIGUSR1];
+            run(val(Sig1));
+         }
+         else if (Signal[SIGUSR2]) {
+            --Signal[0], --Signal[SIGUSR2];
+            run(val(Sig2));
+         }
+         else if (Signal[SIGALRM]) {
+            --Signal[0], --Signal[SIGALRM];
+            run(Alarm);
+         }
+         else if (Signal[SIGINT]) {
+            --Signal[0], --Signal[SIGINT];
+            if (Repl < 2)
+               brkLoad(ex ?: Nil);
+         }
+         else if (Signal[SIGHUP]) {
+            --Signal[0], --Signal[SIGHUP];
+            run(val(Hup));
+         }
+         else if (Signal[SIGTERM]) {
+            for (flg = NO, i = 0; i < Children; ++i)
+               if (Child[i].pid  &&  kill(Child[i].pid, SIGTERM) == 0)
+                  flg = YES;
+            if (flg)
+               break;
+            Signal[0] = 0,  bye(0);
+         }
+      } while (*Signal);
       Env.protect = 0;
    }
 }
@@ -113,14 +129,14 @@ static void sig(int n) {
    if (TtyPid)
       kill(TtyPid, n);
    else
-      Signal = n;
+      ++Signal[n], ++Signal[0];
 }
 
 static void sigTerm(int n) {
    if (TtyPid)
       kill(TtyPid, n);
    else
-      Signal = SIGTERM;
+      ++Signal[SIGTERM], ++Signal[0];
 }
 
 static void sigChld(int n __attribute__((unused))) {
@@ -187,6 +203,21 @@ any doAlarm(any x) {
    return boxCnt(n);
 }
 
+// (sigio ['cnt [. prg]]) -> cnt | prg
+any doSigio(any ex) {
+   any x;
+   int fd;
+
+   if (!isCell(x = cdr(ex)))
+      return Sigio;
+   x = EVAL(car(x)),  fd = (int)xCnt(ex,x);
+   if (isCell(Sigio = cddr(ex))) {
+      fcntl(fd, F_SETOWN, unBox(val(Pid)));
+      fcntl(fd, F_SETFL, fcntl(fd, F_GETFL, 0) | O_NONBLOCK|O_ASYNC);
+   }
+   return x;
+}
+
 // (protect . prg) -> any
 any doProtect(any x) {
    ++Env.protect;
@@ -230,6 +261,15 @@ any doHeap(any x) {
    for (x = Avail;  x;  x = car(x))
       ++n;
    return boxCnt(n / CELLS);
+}
+
+// (adr 'var) -> num
+// (adr 'num) -> var
+any doAdr(any x) {
+   x = cdr(x);
+   if (isNum(x = EVAL(car(x))))
+      return (any)(unDig(x) * WORD);
+   return box(num(x) / WORD);
 }
 
 // (env ['lst] | ['sym 'val] ..) -> lst
@@ -437,7 +477,7 @@ void err(any ex, any x, char *fmt, ...) {
             }
    }
    Chr = ExtN = 0;
-   Env.brk = NO;
+   Break = NO;
    Alarm = Line = Nil;
    f.pid = 0,  f.fd = STDERR_FILENO,  pushOutFiles(&f);
    if (InFile && InFile->name) {
@@ -651,7 +691,11 @@ bool sharedLib(any x) {
    *p++ = '\0';
    {
       int n = Home? strlen(Home) : 0;
+#ifndef __CYGWIN__
       char buf[n + strlen(nm) + 4 + 1];
+#else
+      char buf[n + strlen(nm) + 4 + 4 + 1];
+#endif
 
       if (strchr(nm,'/'))
          strcpy(buf, nm);
@@ -659,6 +703,9 @@ bool sharedLib(any x) {
          if (n)
             memcpy(buf, Home, n);
          strcpy(buf + n, "lib/"),  strcpy(buf + n + 4, nm);
+#ifdef __CYGWIN__
+         strcpy(buf + n + 4 + strlen(nm), ".dll");
+#endif
       }
       if (!(h = dlopen(buf, RTLD_LAZY | RTLD_GLOBAL))  ||  !(h = dlsym(h,p)))
          return NO;
@@ -684,7 +731,7 @@ static any evList2(any foo, any ex) {
    for (;;) {
       if (isNil(val(foo)))
          undefined(foo,ex);
-      if (Signal)
+      if (*Signal)
          sighandler(ex);
       if (isNum(foo = val(foo))) {
          foo = evSubr(foo,ex);
@@ -706,7 +753,7 @@ any evList(any ex) {
    if (!isSym(foo = car(ex))) {
       if (isNum(foo))
          return ex;
-      if (Signal)
+      if (*Signal)
          sighandler(ex);
       if (isNum(foo = evList(foo)))
          return evSubr(foo,ex);
@@ -715,7 +762,7 @@ any evList(any ex) {
    for (;;) {
       if (isNil(val(foo)))
          undefined(foo,ex);
-      if (Signal)
+      if (*Signal)
          sighandler(ex);
       if (isNum(foo = val(foo)))
          return evSubr(foo,ex);
@@ -989,32 +1036,33 @@ any doFile(any ex __attribute__((unused))) {
    return Pop(c1);
 }
 
-// (dir ['any]) -> lst
+// (dir ['any] ['flg]) -> lst
 any doDir(any x) {
    any y;
    DIR *dp;
    struct dirent *p;
    cell c1;
 
-   if (isNil(x = evSym(cdr(x))))
+   if (isNil(y = evSym(x = cdr(x))))
       dp = opendir(".");
    else {
-      char nm[pathSize(x)];
+      char nm[pathSize(y)];
 
-      pathString(x, nm);
+      pathString(y, nm);
       dp = opendir(nm);
    }
    if (!dp)
       return Nil;
+   x = cdr(x),  x = EVAL(car(x));
    do {
       if (!(p = readdir(dp))) {
          closedir(dp);
          return Nil;
       }
-   } while (p->d_name[0] == '.');
+   } while (isNil(x) && p->d_name[0] == '.');
    Push(c1, y = cons(mkStr(p->d_name), Nil));
    while (p = readdir(dp))
-      if (p->d_name[0] != '.')
+      if (!isNil(x) || p->d_name[0] != '.')
          y = cdr(y) = cons(mkStr(p->d_name), Nil);
    closedir(dp);
    return Pop(c1);
@@ -1100,7 +1148,7 @@ static void init(int ac, char *av[]) {
    Env.put = putStdout;
    initOutFile(STDERR_FILENO);
    OutFile = initOutFile(STDOUT_FILENO);
-   Env.task = Alarm = Line = Nil;
+   Env.task = Alarm = Sigio = Line = Nil;
    setrlimit(RLIMIT_STACK, &ULim);
    Tio = tcgetattr(STDIN_FILENO, &OrgTermio) == 0;
    ApplyArgs = cons(cons(consSym(Nil,Nil), Nil), Nil);
@@ -1113,6 +1161,7 @@ static void init(int ac, char *av[]) {
    iSignal(SIGUSR2, sig);
    iSignal(SIGALRM, sig);
    iSignal(SIGTERM, sig);
+   iSignal(SIGIO, sig);
    signal(SIGCHLD, sigChld);
    signal(SIGPIPE, SIG_IGN);
    signal(SIGTTIN, SIG_IGN);
